+++ packages/base-editor/src/commands/keymap.keydown.js
@@ -1,0 +1,10 @@
+export default {
+  command: "keymap.keydown",
+  execute: function (editor, e) {
+    editor.keyboard.add(e.code, e.keyCode, e);
+
+    if (editor.shortcuts) {
+      editor.shortcuts.execute(e, "keydown");
+    }
+  },
+};