+++ packages/base-editor/src/commands/keymap.keyup.js
@@ -1,0 +1,8 @@
+export default {
+  command: "keymap.keyup",
+  execute: function (editor, e) {
+    if (editor.shortcuts) {
+      editor.shortcuts.execute(e, "keyup");
+    }
+  },
+};