+++ packages/base-editor/src/components/editor/BaseEditor.jsx
@@ -1,7 +1,13 @@
-import { classnames, useComponentRender } from "@elf-framework/sapa";
+import {
+  classnames,
+  IF,
+  KEYDOWN,
+  KEYUP,
+  useComponentRender,
+} from "@elf-framework/sapa";
 
 import { Editor, useEditor } from "./Editor";
-
+const formElements = ["TEXTAREA", "INPUT", "SELECT"];
 export class BaseEditor extends Editor {
   template() {
     useComponentRender("editor.plugin.activated");
@@ -19,5 +25,22 @@ export class BaseEditor extends Editor {
         {editor.isPluginActivated ? editor.getUIList("renderView") : undefined}
       </div>
     );
+  }
+
+  isNotFormElement(e) {
+    var tagName = e.target.tagName;
+
+    if (formElements.includes(tagName)) return false;
+    else if (e.target.getAttribute("contenteditable") === "true") return false;
+
+    return true;
+  }
+
+  [KEYDOWN("document") + IF("isNotFormElement")](e) {
+    this.$editor.commands.execute("keymap.keydown", e);
+  }
+
+  [KEYUP("document") + IF("isNotFormElement")](e) {
+    this.$editor.commands.execute("keymap.keyup", e);
   }
 }